% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_rates}
\alias{calc_rates}
\title{Calculate rates from ipums data and associated geography}
\usage{
calc_rates(x, geo_dir)
}
\description{
Calculate rates from ipums data and associated geography
}
\examples{
\dontrun{
calc_rates(dd, geo_dir = "ext-data/idhs/geo")
}
}
