% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_geo_scatter.R
\name{geo_scatter_vis}
\alias{geo_scatter_vis}
\title{Create linked scatter and geo visualization}
\usage{
geo_scatter_vis(rates, title, subtitle, ylab, geo_title, geo_dir, bins = NULL)
}
\arguments{
\item{rates}{TODO}

\item{title}{TODO}

\item{geo_dir}{TODO}

\item{bins}{TODO}
}
\description{
Create linked scatter and geo visualization
}
