% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_geo.R
\name{geo_vis_all}
\alias{geo_vis_all}
\title{Create Mapbox visualization of rate data}
\usage{
geo_vis_all(
  rates,
  geo_dir,
  title = "",
  bins = NULL,
  width = "100vw",
  height = "100vh"
)
}
\arguments{
\item{rates}{TODO}

\item{geo_dir}{TODO}

\item{title}{TODO}

\item{bins}{TODO}

\item{width}{TODO}

\item{height}{TODO}
}
\description{
Create Mapbox visualization of rate data
}
