% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_geo}
\alias{preprocess_geo}
\title{Preprocess (download) shapfiles associated with an IPUMS extract}
\usage{
preprocess_geo(x, output_dir)
}
\description{
Preprocess (download) shapfiles associated with an IPUMS extract
}
\examples{
\dontrun{
dd <- preprocess_ipums(
  input_xml = "ext-data/idhs/extract2/idhs_00002.xml",
  output_file = "ext-data/idhs/extract2/dd.rds"
)
preprocess_geo(dd, output_dir = "ext-data/idhs/geo")
}
}
