% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_scatter.R
\name{scatter_vis_all}
\alias{scatter_vis_all}
\title{Create scatterplot visualization for all geographies}
\usage{
scatter_vis_all(rates)
}
\arguments{
\item{rates}{TODO}
}
\description{
Create scatterplot visualization for all geographies
}
