% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{view_var_descs}
\alias{view_var_descs}
\title{View html table of IPUMS extract variable descriptions}
\usage{
view_var_descs(x, include_geo = FALSE)
}
\arguments{
\item{x}{TODO}

\item{include_geo}{TODO}
}
\description{
View html table of IPUMS extract variable descriptions
}
