% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_rates}
\alias{calc_rates}
\title{Calculate rates from IPUMS DHS data}
\usage{
calc_rates(x, geo_dir, num_var, num_cond, denom_var = NULL, denom_cond = NULL)
}
\arguments{
\item{x}{TODO}

\item{geo_dir}{TODO}

\item{num_var}{TODO}

\item{num_cond}{TODO}

\item{denom_var}{TODO}

\item{denom_cond}{TODO
Calculate rates from ipums data and associated geography}
}
\description{
Calculate rates from IPUMS DHS data
}
\examples{
\dontrun{
calc_rates(dd, geo_dir = "ext-data/idhs/geo")
}
}
